Published

+++ main.cpp
@@ -6,66 +6,41 @@
 using namespace std;
 using json = nlohmann::json;
 
-// MIT License
+// deleted snippets are not converted despite being in the JSON file
 
-// Copyright (c) 2024 Tomas Mark (tomas@digitalspace.name)
+// if you experiment with the code, you may want to change this to false and back to true to avoid deleting the output folder
+#define ERASE_OUTPUT_DIR_AT_STARTUP true
 
-// Permission is hereby granted, free of charge, to any person obtaining a copy
-// of this software and associated documentation files (the "Software"), to deal
-// in the Software without restriction, including without limitation the rights
-// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-// copies of the Software, and to permit persons to whom the Software is
-// furnished to do so, subject to the following conditions:
-
-// The above copyright notice and this permission notice shall be included in all
-// copies or substantial portions of the Software.
-
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-// SOFTWARE.
-
-// masscode2md
-// This project is for an export the massCode snippets to the MarkDown files
-// each exported MarkDown file is located within the same folder as in MassCode defined
-// Dependencies
-// C++ 17
-// Configuration
-// Set the path to the MassCode JSON file
-// std::string jsonFilePath = "/home/tomas/projects/cpp/masscode2md/db.json";
-
-// Set Output directory
-// std::string outputDir = "/home/tomas/joplin-output/";
-
-// Why?
-// I like Joplin 😉
-// That's all!
-
-int main()
+int main(int argc, char *argv[], char *env[])
 {
-    printf("Hello, from masscode2md!\n");
+    std::cout << "masscode2md" << std::endl;
 
-    // Set the Path to the MassCode JSON file
-    std::string jsonFilePath = "/home/tomas/projects/cpp/masscode2md/db.json";
-    
-    // Set Output directory
-    std::string outputDir = "/home/tomas/joplin-output/";
-
-    // Remove the output directory if it exists
-    if (std::filesystem::exists(outputDir))
+    if (argc != 3)
     {
-        printf("Removing output directory: %s\n", outputDir.c_str());
-        std::filesystem::remove_all(outputDir);
+        std::cout << "Usage: " << argv[0] << " <input_json_path> <output_directory>\n";
+        return 1;
     }
 
-    // Load the JSON file
+    // Set the path to the JSON file
+    std::string jsonFilePath = argv[1];
+
+    // Set the output directory
+    std::string outputDir = argv[2];
+
+    if (ERASE_OUTPUT_DIR_AT_STARTUP)
+    {
+        if (std::filesystem::exists(outputDir))
+        {
+            std::cout << "ERASE_OUTPUT_DIR_AT_STARTUP is true\n";
+            std::cout << "Output directory " << outputDir << "exists. Removing it." << std::endl;
+            std::filesystem::remove_all(outputDir);
+        }
+    }
+
     std::ifstream jsonFile(jsonFilePath);
     if (!jsonFile.is_open())
     {
-        printf("Failed to open JSON file: %s\n", jsonFilePath.c_str());
+        std::cout << "Failed to open JSON file: " << jsonFilePath << std::endl;
         return 1;
     }
 
@@ -77,7 +52,7 @@ int main()
     }
     catch (const std::exception &e)
     {
-        printf("Failed to parse JSON data: %s\n", e.what());
+        std::cout << "Failed to parse JSON data: " << e.what() << std::endl;
         return 1;
     }
 
@@ -86,7 +61,7 @@ int main()
     {
         if (!std::filesystem::is_empty(outputDir))
         {
-            printf("Output folder is not empty. Abort\n");
+            std::cout << "Output folder is not empty. Abort" << std::endl;
             return 1;
         }
     }
@@ -96,13 +71,10 @@ int main()
         // Create the output folder
         if (!std::filesystem::create_directory(outputDir))
         {
-            printf("Failed to create output folder. Abort\n");
+            std::cout << "Failed to create output folder. Abort" << std::endl;
             return 1;
         }
     }
-
-    // Process the JSON data
-    // ...
 
     // Create Folders
     std::vector<std::pair<std::string, std::string>> folderData;
@@ -113,17 +85,16 @@ int main()
         folderData.push_back(std::make_pair(folderName, folderId));
 
         // Create the folder using the folderName
-        // ...
-        printf("Folder: %s, ID: %s\n", folderName.c_str(), folderId.c_str());
+        std::cout << "Folder Name: " << folderName << ", Folder ID: " << folderId << std::endl;
 
-        // Replace invalid characters in folderName with a valid character
+        // Replace invalid characters in label with a valid character
         std::replace_if(folderName.begin(), folderName.end(), [](char c)
                         { return !std::isalnum(c) && c != '_' && c != '-' && c != '+'; }, '_');
 
         std::string folderPath = outputDir + "/" + folderName;
         if (!std::filesystem::create_directory(folderPath))
         {
-            printf("Failed to create folder: %s\n", folderPath.c_str());
+            std::cout << "Failed to create folder: " << folderPath << std::endl;
             return 1;
         }
 
@@ -134,17 +105,18 @@ int main()
     // Process the Snippets
     for (const auto &snippet : jsonData["snippets"])
     {
-        bool isDeleted = snippet["isDeleted"];
+
+        // skip deleted snippets
+        if (snippet["isDeleted"])
+        {
+            continue;
+        }
+
         std::string folderId = snippet["folderId"];
         std::string name = snippet["name"];
 
         // Print the snippet data
-        printf("Snippet: %s, Folder ID: %s\n", name.c_str(), folderId.c_str());
-
-        if (isDeleted)
-        {
-            continue;
-        }
+        std::cout << "Snippet: " << name << ", Folder ID: " << folderId << std::endl;
 
         // Find the folderName for this folderId
         std::string folderName;
@@ -159,7 +131,7 @@ int main()
 
         if (folderName.empty())
         {
-            printf("Failed to find folder for snippet: %s\n", snippet["label"].get<std::string>().c_str());
+            std::cout << "Failed to find folder for snippet: " << snippet["label"].get<std::string>() << std::endl;
             continue;
         }
 
@@ -174,7 +146,7 @@ int main()
 
             if (label.empty() || value.empty())
             {
-                printf("Invalid content in snippet: %s\n", snippet["name"].get<std::string>().c_str());
+                std::cout << "Invalid content in snippet: " << snippet["name"].get<std::string>() << std::endl;
                 continue;
             }
 
@@ -190,19 +162,16 @@ int main()
 
             std::string filePath = outputDir + "/" + folderName + "/" + name + ".md";
 
-            // Write the value to the file
+            // Write the content to the file
             std::ofstream outFile(filePath);
             if (!outFile)
             {
-                printf("Failed to open file: %s\n", filePath.c_str());
+                std::cout << "Failed to open file: " << filePath << std::endl;
                 continue;
             }
 
-            if (language.empty())
-            {
-                language = "plaintext";
-            }
-            else if (language == "c_cpp")
+            // Change the language to cpp if it is c_cpp to match the markdown language
+            if (language == "c_cpp")
             {
                 language = "cpp";
             }
@@ -211,7 +180,9 @@ int main()
             outFile << value;
             outFile << '\n'
                     << "```" << '\n';
+
             outFile.close();
         }
     }
+    std::cout << "Done" << std::endl;
 }
